feat: CategoriesViewModel 분리진행
